separated query model from create model for transform and load
