+++ backend/modules/extract/domain/exceptions.py
@@ -1,10 +1,10 @@
 class DataValidationError(Exception):
     def __init__(self, message: str, file_name: str):
-        super().__init__()
+        super().__init__(message)
         self.file_name = file_name
 
 
 class FileExtensionNotSupportedError(Exception):
     def __init__(self, message: str, file_extension: str):
-        super().__init__()
+        super().__init__(message)
         self.file_extension = file_extension
