+++ backend/modules/extract/domain/value_objects.py
