+++ backend/modules/load/domain/commands.py
@@ -1,9 +1,9 @@
 from dataclasses import dataclass
 
 from ...common.domain import commands as common_commands
-from ...load.services import queries as load_queries
+from ...transform.domain import value_objects as domain_value_objects
 
 
 @dataclass(frozen=True)
 class SaveData(common_commands.DomainCommand):
-    output_data: list[load_queries.OutputData]
+    output_data: list[domain_value_objects.OutputData]
