+++ backend/modules/load/domain/events.py
@@ -1,9 +1,9 @@
 from dataclasses import dataclass
 
 from ...common.domain import events as common_events
-from ...load.services import queries as load_queries
+from ...transform.domain import value_objects as domain_value_objects
 
 
 @dataclass(frozen=True)
 class DataSaved(common_events.DomainEvent):
-    data: load_queries.OutputData
+    data: domain_value_objects.OutputData
