+++ backend/modules/load/domain/ports/repositories.py
@@ -1,12 +1,12 @@
 from abc import abstractmethod
 
 from ....common.domain import ports as common_ports
-from ...services import queries as load_queries
+from ....transform.domain import value_objects as domain_value_objects
 
 
 class AbstractDataDomainRepository(common_ports.AbstractDomainRepository):
     @abstractmethod
-    def create(self, data: list[load_queries.OutputData]) -> None:
+    def create(self, data: list[domain_value_objects.OutputData]) -> None:
         """
         :param: Data to save.
         """
