+++ backend/modules/transform/domain/commands.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
 from ...common.domain import commands as common_commands
-from ...data.domain import value_objects as data_value_objects
+from ...extract.domain import value_objects as data_value_objects
 
 
 @dataclass(frozen=True)
