+++ backend/modules/transform/domain/events.py
@@ -1,9 +1,9 @@
 from dataclasses import dataclass
 
 from ...common.domain import events as common_events
-from ...load.services import queries as load_queries
+from ..domain import value_objects as domain_value_objects
 
 
 @dataclass(frozen=True)
 class DataTransformed(common_events.DomainEvent):
-    output_data: load_queries.OutputData
+    output_data: domain_value_objects.OutputData
