+++ backend/modules/transform/domain/value_objects.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class OutputData:
+    full_name: str
+    age: int
+    is_satisfied: bool
