+++ backend/tests/unit/test_transform.py
@@ -2,8 +2,8 @@ from pathlib import Path
 
 import pandas as pd
 
-from modules.load.services import queries as load_queries
 from modules.transform.domain import commands as domain_commands
+from modules.transform.domain import value_objects as domain_value_objects
 from modules.transform.services.commands import commands as service_commands
 
 
@@ -14,8 +14,10 @@ def test_transformed_data_contains_fullname_column():
     command = domain_commands.TransformData(input_df)
 
     # When
-    result: list[load_queries.OutputData] = service_commands.transform(command)
+    result: list[domain_value_objects.OutputData] = service_commands.transform(command)
 
     # Then
     assert len(result) == test_dataset_size
-    assert result.count(load_queries.OutputData("Jessica Barnes", 58, False)) == 1
+    assert (
+        result.count(domain_value_objects.OutputData("Jessica Barnes", 58, False)) == 1
+    )
